Move group communication down to jvstm-ispn only
  This commit complements cc88fe2, which first added support for group
  communication in all JVSTM-based backends.  It is considered best for now to
  keep this backend-specific, so, in this commit, we move such functionality
  to the only backend that currently extends jvstm-common: jvsmt-ispn.
