+++ backend/jvstm-common/runtime/pom.xml
@@ -74,10 +74,6 @@
             <groupId>jvstm</groupId>
             <artifactId>jvstm-fenix</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.hazelcast</groupId>
-            <artifactId>hazelcast</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
