+++ backend/jvstm-infinispan/runtime/src/main/java/pt/ist/fenixframework/backend/jvstm/infinispan/JvstmIspnConfig.java
@@ -12,6 +12,11 @@ import org.slf4j.LoggerFactory;
 
 import pt.ist.fenixframework.backend.jvstm.JVSTMConfig;
 
+import com.hazelcast.config.ClasspathXmlConfig;
+//import com.hazelcast.core.AtomicNumber;
+//import com.hazelcast.core.Hazelcast;
+//import com.hazelcast.core.HazelcastInstance;
+
 /**
  * This is the configuration manager used by the fenix-framework-backend-jvstm-infinispan project.
  * 
@@ -22,13 +27,25 @@ public class JvstmIspnConfig extends JVSTMConfig {
 
     private static final String FAILED_INIT = "Failed to initialize Backend JVSTM-Infinispan";
 
+    protected static final String HAZELCAST_FF_GROUP_NAME = "FenixFrameworkGroup";
+
+    /**
+     * This <strong>optional</strong> parameter specifies the Hazelcast configuration file. This
+     * configuration will used to create a group communication system between Fenix Framework nodes. The default value
+     * for this parameter is <code>fenix-framework-hazelcast-default.xml</code>, which is the default
+     * configuration file that ships with the framework.
+     */
+    protected String hazelcastConfigFile = "fenix-framework-hazelcast-default.xml";
+
     /**
      * This <strong>optional</strong> parameter specifies the location of the XML file used to configure Infinispan. This file
      * should be available in the application's classpath.
      */
     protected String ispnConfigFile = null;
 
-    // process this config's parameters
+    public String getHazelcastConfigFile() {
+        return hazelcastConfigFile;
+    }
 
     public String getIspnConfigFile() {
         return this.ispnConfigFile;
@@ -49,6 +66,13 @@ public class JvstmIspnConfig extends JVSTMConfig {
     @Override
     public String getBackEndName() {
         return JvstmIspnBackEnd.BACKEND_NAME;
+    }
+
+    public com.hazelcast.config.Config getHazelcastConfig() {
+        System.setProperty("hazelcast.logging.type", "slf4j");
+        com.hazelcast.config.Config hzlCfg = new ClasspathXmlConfig(getHazelcastConfigFile());
+        hzlCfg.getGroupConfig().setName(HAZELCAST_FF_GROUP_NAME);
+        return hzlCfg;
     }
 
 }
