+++ test/test-backend-jvstm-common/src/test/resources/log4j.properties
@@ -1,7 +1,7 @@
 log4j.logger.test.backend=WARN, FFAPEND
 log4j.logger.pt.ist.fenixframework=WARN, FFAPEND
 
-log4j.logger.org.jgroups=WARN, FFAPEND
+log4j.logger.org.jgroups=ERROR, FFAPEND
 log4j.logger.org.hibernate.search.impl=WARN, FFAPEND
 log4j.logger.org=WARN, FFAPEND
 log4j.logger.com=WARN, FFAPEND
